greatfireoflondonmobileappgame/README.md
The invention on the Great Fire of London Mobile App Game came to Errol Baptiste after visiting various exhibitions and events during the year of 2016 that detailed in depth the significant on the 350th Anniversary of this historical event. It led him of daring to be different by starting the birth of idea, creating good opportunities, change of transformation, process of change and the legacy. Errol felt his brilliant idea on the virtual reality (VR) Great Fire of London Mobile App Trail Game as a game changer by doing things differently over the past 50 years since 1966.